SpatialBernoulli

Documentation for SpatialBernoulli.

SpatialBernoulli.SB

SpatialBernoulli.SB

Distributions.fit_mle

Distributions.pdf

SpatialBernoulli.matern

SpatialBernoulli.SB — Type

SB{TR<:Real, TS<:Real, TO<:Real, AV<:AbstractVector, AM<:AbstractMatrix, AAM<:AbstractMatrix}

Defines a discrete multivariate distribution SB using a latend Gaussian process. The latent covarience matrix is definied by a Matern covariance (range, sill, order). The marginal Bernoulli probabilities are given by λ.

SpatialBernoulli.SB — Method

SB(range, sill, order, λ, h)

Constructor for a discrete multivariate distribution SB using a latend Gaussian process. The latent covarience matrix is definied by a Matern covariance (range, sill, order). The marginal Bernoulli probabilities are given by λ. The constructor used the distance martix to compute the covariance matrix.

Distributions.fit_mle — Method

fit_mle(d::SB, Y::AbstractArray{<:Real}[,wp::AbstractMatrix{<:Real}, w::AbstractVector{<:Real}]; solver = Optim.LBFGS(), m = 1000*length(d), return_sol = false, solkwargs...)

Return the (weigthed) MLE for the distribution d::SB.

solver choice of solver

solkwargs any keywords arguments that can be put inside the solve functions e.g. fit_mle(d, Y; maxiters = 2)

w not necessary, used for using weighted loglikelihood instead of loglikelihood

wp not technically necessary, used for using pairwise loglikelihood instead of true loglikelihood.

Distributions.pdf — Method

pdf(d::SB, y::AbstractVector{<:Real}; m=length(d) * 100, return_error=false)

pdf(y) = CDF(N(mu,sigma))(0,...,0) is all there is to compute ! $h(y) = \int_{-infty}^0 \dots \int_{-infty}^0 pdf(\mathcal{N}((-1)^{y_s}\sigma_s \phi^{-1}(\lambda_s)),\Sigma U)$ recommended to use m= d1000 inside of MvNormalCDF, set to d100 as a starting point for this work (is quite ok)

SpatialBernoulli.matern — Method

matern(h; range, sill, order) expkernel(h;range,sill) Define the matérn and exponential kernel.